+++ tests/CRUDlexUserTestEnv/TestDBSetup.php
@@ -1,43 +1,51 @@
 <?php
 namespace CRUDlexUserTestEnv;
 
+use CRUDlex\EntityDefinitionFactory;
+use CRUDlex\EntityDefinitionValidator;
+use Doctrine\DBAL\Connection;
 use League\Flysystem\Adapter\NullAdapter;
+use League\Flysystem\Filesystem;
 use Silex\Application;
 use Silex\Provider\DoctrineServiceProvider;
 
 use Eloquent\Phony\Phpunit\Phony;
 
 use CRUDlex\MySQLDataFactory;
-use CRUDlex\ServiceProvider;
+use CRUDlex\Service;
 use CRUDlex\UserSetup;
 use CRUDlexUserTestEnv\NullFileProcessor;
+use Symfony\Component\Routing\Generator\UrlGenerator;
+use Symfony\Component\Routing\RequestContext;
+use Symfony\Component\Routing\RouteCollection;
+use Symfony\Component\Translation\Translator;
 
 class TestDBSetup {
 
     private static $filesystemHandle;
 
-    public static function createAppAndDB($useManyToMany)
+    public static function getDBConfig()
     {
-        $app = new Application();
-        $app->register(new DoctrineServiceProvider(), [
-            'dbs.options' => [
-                'default' => [
-                    'host'      => '127.0.0.1',
-                    'dbname'    => 'crudTest',
-                    'user'      => 'root',
-                    'password'  => '',
-                    'charset'   => 'utf8',
-                ]
-            ],
-        ]);
+        return [
+            'host'      => '127.0.0.1',
+            'dbname'    => 'crudTest',
+            'user'      => 'root',
+            'password'  => '',
+            'charset'   => 'utf8',
+            'driver' => 'pdo_mysql',
+        ];
+    }
 
-        $app['db']->executeUpdate('DROP TABLE IF EXISTS password_reset;');
-        $app['db']->executeUpdate('DROP TABLE IF EXISTS user_role;');
-        $app['db']->executeUpdate('DROP TABLE IF EXISTS role;');
-        $app['db']->executeUpdate('DROP TABLE IF EXISTS user;');
+    public static function createDB(Connection $db, $useManyToMany)
+    {
 
-        $app['db']->executeUpdate('SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";');
-        $app['db']->executeUpdate('SET time_zone = "+00:00"');
+        $db->executeUpdate('DROP TABLE IF EXISTS password_reset;');
+        $db->executeUpdate('DROP TABLE IF EXISTS user_role;');
+        $db->executeUpdate('DROP TABLE IF EXISTS role;');
+        $db->executeUpdate('DROP TABLE IF EXISTS user;');
+
+        $db->executeUpdate('SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";');
+        $db->executeUpdate('SET time_zone = "+00:00"');
 
         $sql = 'CREATE TABLE `role` ('.
             '  `id` int(11) NOT NULL AUTO_INCREMENT,'.
@@ -48,7 +56,7 @@ class TestDBSetup {
             '  `role` varchar(255) NOT NULL,'.
             '  PRIMARY KEY (`id`)'.
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8;';
-        $app['db']->executeUpdate($sql);
+        $db->executeUpdate($sql);
 
         $sql = 'CREATE TABLE `user` ('.
             '  `id` int(11) NOT NULL AUTO_INCREMENT,'.
@@ -62,7 +70,7 @@ class TestDBSetup {
             '  `email` varchar(255) NOT NULL,'.
             '  PRIMARY KEY (`id`)'.
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8;';
-        $app['db']->executeUpdate($sql);
+        $db->executeUpdate($sql);
 
         if ($useManyToMany) {
             $sql = 'CREATE TABLE `user_role` ('.
@@ -73,7 +81,7 @@ class TestDBSetup {
                 '  CONSTRAINT `userrole_ibfk_1` FOREIGN KEY (`user`) REFERENCES `user` (`id`),'.
                 '  CONSTRAINT `userrole_ibfk_2` FOREIGN KEY (`role`) REFERENCES `role` (`id`)'.
                 ') ENGINE=InnoDB DEFAULT CHARSET=utf8;';
-            $app['db']->executeUpdate($sql);
+            $db->executeUpdate($sql);
         } else {
             $sql = 'CREATE TABLE `user_role` ('.
                 '  `id` int(11) NOT NULL AUTO_INCREMENT,'.
@@ -89,7 +97,7 @@ class TestDBSetup {
                 '  CONSTRAINT `userrole_ibfk_1` FOREIGN KEY (`user`) REFERENCES `user` (`id`),'.
                 '  CONSTRAINT `userrole_ibfk_2` FOREIGN KEY (`role`) REFERENCES `role` (`id`)'.
                 ') ENGINE=InnoDB DEFAULT CHARSET=utf8;';
-            $app['db']->executeUpdate($sql);
+            $db->executeUpdate($sql);
         }
 
         $sql = 'CREATE TABLE `password_reset` ('.
@@ -105,31 +113,37 @@ class TestDBSetup {
             '  KEY `user` (`user`),'.
             '  CONSTRAINT `passwordreset_ibfk_1` FOREIGN KEY (`user`) REFERENCES `user` (`id`)'.
             ') ENGINE=InnoDB DEFAULT CHARSET=utf8;';
-        $app['db']->executeUpdate($sql);
-        return $app;
+        $db->executeUpdate($sql);
     }
 
-    public static function createServiceProvider($useManyToMany)
+    public static function createService($useManyToMany)
     {
+
+
+        $config = new \Doctrine\DBAL\Configuration();
+        $db = \Doctrine\DBAL\DriverManager::getConnection(TestDBSetup::getDBConfig(), $config);
+        static::createDB($db, $useManyToMany);
+        $crudFile =  __DIR__.'/../'.($useManyToMany ? 'crudManyToMany.yml' : 'crud.yml');
+        $dataFactory = new MySQLDataFactory($db);
+        $filesystem = new Filesystem(new NullAdapter());
+        $validator = new EntityDefinitionValidator();
+        $routes = new RouteCollection();
+        $context = new RequestContext();
+        $urlGenerator = new UrlGenerator($routes, $context);
+
+        $translator = new Translator('en');
+        $entityDefinitionFactory = new EntityDefinitionFactory();
+        $service = new Service($crudFile, null, $urlGenerator, $translator, $dataFactory, $entityDefinitionFactory, $filesystem, $validator);
 
         static::$filesystemHandle = Phony::partialMock('\\League\\Flysystem\\Filesystem', [new NullAdapter()]);
         static::$filesystemHandle->readStream->returns(null);
         static::$filesystemHandle->getMimetype->returns('test');
         static::$filesystemHandle->getSize->returns(42);
 
-        $app = self::createAppAndDB($useManyToMany);
-        $crudServiceProvider = new ServiceProvider();
-
-        $app['crud.filesystem'] = static::$filesystemHandle->get();
-        $app['crud.datafactory'] = new MySQLDataFactory($app['db']);
-        $app['crud.file'] = __DIR__.'/../'.($useManyToMany ? 'crudManyToMany.yml' : 'crud.yml');
-        $crudServiceProvider->boot($app);
-        $crudServiceProvider->init(null, $app);
-
         $userSetup = new UserSetup();
-        $userSetup->addEvents($crudServiceProvider->getData('user'));
+        $userSetup->addEvents($service->getData('user'));
 
-        return $crudServiceProvider;
+        return $service;
     }
 
     public static function getFileProcessor()